+++ boyar/config/node_configuration_test.go
@@ -1,7 +1,7 @@
 package config
 
 import (
-	"github.com/orbs-network/boyarin/strelets"
+	"github.com/orbs-network/boyarin/boyar/topology"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
@@ -15,7 +15,7 @@ func TestNodeConfigurationContainer_Hash(t *testing.T) {
 
 	require.EqualValues(t, oldHash, source.Hash(), "hash can't change if the value didn't change")
 
-	source.value.FederationNodes = []*strelets.FederationNode{
+	source.value.FederationNodes = []*topology.FederationNode{
 		{IP: "1.2.3.4", Address: "some-fake-key"},
 	}
