+++ boyar/provision_virtual_chain.go
@@ -117,3 +117,35 @@ func getNetworkConfigJSON(nodes []*topology.FederationNode) []byte {
 
 	return json
 }
+
+func buildPeersMap(nodes []*topology.FederationNode, gossipPort int) *config.PeersMap {
+	peersMap := make(config.PeersMap)
+
+	for _, node := range nodes {
+		// Need this override for more flexibility in network config and also for local testing
+		port := node.Port
+		if port == 0 {
+			port = gossipPort
+		}
+
+		peersMap[config.NodeAddress(node.Address)] = &config.Peer{
+			node.IP, port,
+		}
+	}
+
+	return &peersMap
+}
+
+func getVirtualChainConfig(cfg config.NodeConfiguration, chain *config.VirtualChain) *config.VirtualChainConfig {
+	peers := buildPeersMap(cfg.FederationNodes(), chain.GossipPort)
+
+	signerOn := cfg.Services().SignerOn()
+	keyPairConfig := getKeyConfigJson(cfg, signerOn)
+
+	return &config.VirtualChainConfig{
+		VirtualChain:  chain,
+		Peers:         peers,
+		NodeAddress:   cfg.NodeAddress(),
+		KeyPairConfig: keyPairConfig,
+	}
+}
