+++ boyar/provision_virtual_chain_test.go
@@ -1,20 +1,22 @@
-package config
+package boyar
 
 import (
 	"fmt"
+	"github.com/orbs-network/boyarin/boyar/topology"
 	"github.com/orbs-network/boyarin/test/helpers"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
 
 func Test_getNetworkConfigJSON(t *testing.T) {
-	peers := make(PeersMap)
+	var nodes []*topology.FederationNode
 
-	for i, key := range helpers.NodeAddresses() {
-		peers[NodeAddress(key)] = &Peer{
-			IP:   fmt.Sprintf("10.0.0.%d", i+1),
-			Port: 4400 + i,
-		}
+	for i, address := range helpers.NodeAddresses() {
+		nodes = append(nodes, &topology.FederationNode{
+			Address: address,
+			IP:      fmt.Sprintf("10.0.0.%d", i+1),
+			Port:    4400 + i,
+		})
 	}
 
 	require.JSONEq(t, `{
@@ -24,5 +26,5 @@ func Test_getNetworkConfigJSON(t *testing.T) {
 			{"address":"a328846cd5b4979d68a8c58a9bdfeee657b34de7","ip":"10.0.0.1","port":4400},
 			{"address":"6e2cb55e4cbe97bf5b1e731d51cc2c285d83cbf9","ip":"10.0.0.3","port":4402}
 		]
-	}`, string(getNetworkConfigJSON(&peers)))
+	}`, string(getNetworkConfigJSON(nodes)))
 }
